*** По мелочи: *** (+) Кальуляция веса оружия (учитывая тип патрона и гранаты) //by Sin! // (~) Немного поправил предупреждения из-за "libtheora" (+) Включил правку от Real Wolf: Уменьшаем вес и цену предмета, после использования. (-) Удалены некоторые маркеры типа 'bla-bla-bla_shkiper_marker' (~) Убрано отображение "скорострельности, точности и т.д." для ножа. (проверка по классу предмета) (~) [Bugfix] Crows fix // by Sin! //

+++ 3rd party/Xiph/libtheora/lib/mcenc.c
@@ -629,7 +629,7 @@ static unsigned oc_mcenc_ysatd_halfpel_mbrefine(const oc_enc_ctx *_enc,
     mvoffset1=mvoffset_base+(dx&~xmask)+(offset_y[site]&~ymask);
     err=oc_satd16_halfpel(_enc,frag_buf_offs,fragis,
      mvoffset0,mvoffset1,src,ref,ystride,_best_err);
-    if(err<_best_err){
+    if(err<(int)_best_err){
       _best_err=err;
       best_site=site;
     }
