+++ 3rd party/Xiph/libtheora/lib/state.c
@@ -506,7 +506,7 @@ static int oc_state_ref_bufs_init(oc_theora_state *_state,int _nrefs){
   ref_frame_data_sz=_nrefs*ref_frame_sz;
   /*Check for overflow.
     The same caveats apply as for oc_state_frarray_init().*/
-  if(yplane_sz/yhstride!=yheight||2*cplane_sz<cplane_sz||
+  if((int)yplane_sz/yhstride!=yheight||2*cplane_sz<cplane_sz||
    ref_frame_sz<yplane_sz||ref_frame_data_sz/_nrefs!=ref_frame_sz){
     return TH_EIMPL;
   }
